Almost finish : Remainig: change DTO paths + add Logger

+++ Ecommerce.Application/Services/CategoryServices.cs
@@ -1,29 +1,29 @@
-﻿using Ecommerce.Application.DTOs;
+﻿using System.ComponentModel.DataAnnotations;
+using Ecommerce.Application.DTOs;
 using Ecommerce.Application.Interfaces;
 using Ecommerce.Core.Entities;
 using Ecommerce.Core.Interfaces;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.EntityFrameworkCore.Query.SqlExpressions;
 
 namespace Ecommerce.Application.Services;
 
 public class CategoryServices : ICategoryService
 {
+    public const string CategoryException = "Category Not Found!";
+    public const string ParentCategoryException = "Parent Category Not Found!";
     private readonly ICategoryRepository _categoryRepository;
 
     public CategoryServices(ICategoryRepository categoryRepository)
     {
-        _categoryRepository = categoryRepository;
+        this._categoryRepository = categoryRepository;
     }
-
     public async Task<CategoryDto> AddCategoryAsync(AddUpdateCategoryDto updateCategoryDto)
     {
-        if (updateCategoryDto.ParentCategoryName == " ")
+        if (updateCategoryDto.ParentName == " ")
         {
             var categoryParent = new Category
             {
                 Id = Guid.NewGuid(),
-                Name = updateCategoryDto.CategoryName,
+                Name = updateCategoryDto.Name,
                 Type = updateCategoryDto.Type,
                 SubCategories = []
             };
@@ -38,15 +38,23 @@ public class CategoryServices : ICategoryService
             };
         }
 
-        var parentCategory = await _categoryRepository.GetCategoryByName(updateCategoryDto.ParentCategoryName);
+        var parentCategory = await _categoryRepository.GetCategoryByName(updateCategoryDto.ParentName);
+        if (parentCategory == null)
+        {
+            throw new Exception(CategoryException);
+        }
         var childCategory = new Category
         {
             Id = Guid.NewGuid(),
-            Name = updateCategoryDto.CategoryName,
+            Name = updateCategoryDto.Name,
             ParentCategoryId = parentCategory.Id,
             Type = updateCategoryDto.Type,
+            SubCategories = []
         };
+
+      
         await _categoryRepository.AddNewCategory(childCategory);
+        await _categoryRepository.AddSubCategoryAsync(parentCategory, childCategory);
         return new CategoryDto
         {
             Id = childCategory.Id,
@@ -73,6 +81,10 @@ public class CategoryServices : ICategoryService
     public async Task<CategoryDto> GetCategoryByNameAsync(string name)
     {
         var cat = await _categoryRepository.GetCategoryByName(name);
+        if (cat == null)
+        {
+            throw new Exception(CategoryException);
+        }
         if (!cat.Type)
         {
             return new CategoryDto
@@ -81,6 +93,7 @@ public class CategoryServices : ICategoryService
                 CategoryName = cat.Name,
                 ParentCategoryId = cat.ParentCategoryId,
                 Type = cat.Type,
+                SubCategories = cat.SubCategories
             };
         }
 
@@ -90,26 +103,40 @@ public class CategoryServices : ICategoryService
             CategoryName = cat.Name,
             ParentCategoryId = cat.ParentCategoryId,
             Type = cat.Type,
+            SubCategories = cat.SubCategories
         };
     }
 
     public async Task<CategoryDto> GetParentCategoryAsync(Guid childId)
     {
         var parent = await _categoryRepository.GetParentByChildId(childId);
+        if (parent == null)
+        {
+            throw new Exception(ParentCategoryException);
+        }
+
+        if (!parent.Type)
+        {
+            throw new Exception("There is no subcategory for a subcategory");
+        }
         var child = await _categoryRepository.GetCategoryById(childId);
         return new CategoryDto
         {
             Id = parent.Id,
             CategoryName = parent.Name,
-            ParentCategoryId = (Guid)child.ParentCategoryId!,
+            ParentCategoryId = child.ParentCategoryId,
             Type = parent.Type,
+            SubCategories = parent.SubCategories
         };
     }
 
     public async Task<CategoryDto> GetCategoryByIdAsync(Guid id)
     {
         var cat = await _categoryRepository.GetCategoryById(id);
-        var cats = await _categoryRepository.GetAllCategories();
+        if (cat == null)
+        {
+            throw new Exception(CategoryException);
+        }
         return new CategoryDto
         {
             Id = cat.Id,
@@ -125,12 +152,12 @@ public class CategoryServices : ICategoryService
         var targetCategory = await _categoryRepository.GetCategoryById(id);
         if (targetCategory == null)
         {
-            return null;
+            throw new Exception(CategoryException);
         }
-        targetCategory.Name = updateCategoryDto.CategoryName;
+        targetCategory.Name = updateCategoryDto.Name;
         targetCategory.Type = updateCategoryDto.Type;
        
-        if (updateCategoryDto.ParentCategoryName == " ")
+        if (updateCategoryDto.ParentName == " ")
         {
             await _categoryRepository.UpdateCategory(targetCategory);
             return new CategoryDto
@@ -139,11 +166,17 @@ public class CategoryServices : ICategoryService
                 ParentCategoryId =new Guid("00000000-0000-0000-0000-000000000000"),
                 ParentCategoryName = "This is a parent",
                 Type = targetCategory.Type,
-                CategoryName = targetCategory.Name
+                CategoryName = targetCategory.Name,
+                SubCategories = targetCategory.SubCategories
             };
         }
-        var newParentCat = await _categoryRepository.GetCategoryByName(updateCategoryDto.ParentCategoryName);
+        var newParentCat = await _categoryRepository.GetCategoryByName(updateCategoryDto.ParentName);
+        if (newParentCat == null)
+        {
+            throw new Exception(ParentCategoryException);
+        }
         targetCategory.ParentCategoryId = newParentCat.Id;
+        await _categoryRepository.AddSubCategoryAsync(newParentCat, targetCategory);
        
         await _categoryRepository.UpdateCategory(targetCategory);
         return new CategoryDto
@@ -152,39 +185,31 @@ public class CategoryServices : ICategoryService
             ParentCategoryId = targetCategory.ParentCategoryId,
             ParentCategoryName = newParentCat.Name,
             Type = targetCategory.Type,
-            CategoryName = targetCategory.Name
+            CategoryName = targetCategory.Name,
+            SubCategories = targetCategory.SubCategories
         };
     }
 
     public async Task<bool> DeleteCategoryByIdAsync(Guid id)
     {
         var category = await _categoryRepository.GetCategoryById(id);
+        if (category == null)
+        {
+            throw new Exception(CategoryException);
+        }
         if (!category.Type)
         {
             await _categoryRepository.DeleteCategoryById(id);
             return true;
         }
-
-
-        var subCategories = await _categoryRepository.GetSubCategory(id);
-        if (subCategories.Count() != 0)
+        
+        if (category.SubCategories.Count() != 0)
         {
-            return false;
+            throw new ValidationException("You cannot delete this category before its children!");
         }
 
         await _categoryRepository.DeleteCategoryById(id);
         return true;
     }
-
-    public async Task<IEnumerable<CategoryDto>> GetSubCategoriesAsync(Guid parentCategory)
-    {
-        var categories = await _categoryRepository.GetSubCategory(parentCategory);
-        return categories.Select(cat => new CategoryDto
-        {
-            Id = cat.Id,
-            CategoryName = cat.Name,
-            ParentCategoryId = (Guid)cat.ParentCategoryId!,
-            Type = cat.Type
-        });
-    }
+    
 }
